Move TR computation to TurmoilHandler, refactor to only accept TRSource.
