+++ src/Player.ts
@@ -13,7 +13,7 @@ import {CorporationCard} from './cards/corporation/CorporationCard';
 import {Game} from './Game';
 import {HowToPay} from './inputs/HowToPay';
 import {IAward} from './awards/IAward';
-import {ICard, IResourceCard} from './cards/ICard';
+import {ICard, IResourceCard, TRSource} from './cards/ICard';
 import {Colony} from './colonies/Colony';
 import {ISerializable} from './ISerializable';
 import {IMilestone} from './milestones/IMilestone';
@@ -1800,7 +1800,6 @@ export class Player implements ISerializable<SerializedPlayer> {
 
   public canPlay(card: IProjectCard): boolean {
     const baseCost = this.getCardCost(card);
-    const redsCost = this.computeTerraformRatingBump(card) * REDS_RULING_POLICY_COST;
 
     const canAfford = this.canAfford(
       baseCost,
@@ -1811,7 +1810,7 @@ export class Player implements ISerializable<SerializedPlayer> {
         microbes: this.canUseMicrobes(card),
         science: this.canUseScience(card),
         reserveUnits: MoonExpansion.adjustedReserveCosts(this, card),
-        redsCost,
+        tr: card.tr,
       });
 
     if (!canAfford) {
@@ -1839,7 +1838,7 @@ export class Player implements ISerializable<SerializedPlayer> {
     microbes?: boolean,
     science?: boolean,
     reserveUnits?: Units,
-    redsCost?: number,
+    tr?: TRSource,
   }) {
     const reserveUnits = options?.reserveUnits ?? Units.EMPTY;
     if (!this.hasUnits(reserveUnits)) {
@@ -1852,7 +1851,9 @@ export class Player implements ISerializable<SerializedPlayer> {
     const canUseMicrobes: boolean = options?.microbes ?? false;
     const canUseScience: boolean = options?.science ?? false;
 
-    const availableMegacredits = this.megaCredits - (reserveUnits.megacredits + (options?.redsCost ?? 0));
+    const redsCost = TurmoilHandler.computeTerraformRatingBump(this, options?.tr) * REDS_RULING_POLICY_COST;
+
+    const availableMegacredits = this.megaCredits - (reserveUnits.megacredits + redsCost);
     if (availableMegacredits < 0) {
       return false;
     }
@@ -1863,74 +1864,6 @@ export class Player implements ISerializable<SerializedPlayer> {
       (canUseFloaters ? this.getFloatersCanSpend() * 3 : 0) +
       (canUseMicrobes ? this.getMicrobesCanSpend() * 2 : 0) +
       (canUseScience ? this.getSpendableScienceResources() : 0);
-  }
-
-  // TODO(kberg): Move this to somewhere in turmoil/
-  // TODO(kberg): Add a test where if you raise oxygen to max temperature but temperature is maxed you do not have to pay for it.
-  // It works, but4 a test would be helpful.
-  private computeTerraformRatingBump(card: IProjectCard): number {
-    if (!PartyHooks.shouldApplyPolicy(this, PartyName.REDS)) return 0;
-
-    let tr = card.tr;
-    if (tr === undefined) return 0;
-
-    // Local copy
-    tr = {...tr};
-    let total = 0;
-
-    if (tr.oxygen !== undefined) {
-      const availableSteps = constants.MAX_OXYGEN_LEVEL - this.game.getOxygenLevel();
-      const steps = Math.min(availableSteps, tr.oxygen);
-      total = total + steps;
-      // TODO(kberg): Add constants for these constraints.
-      if (this.game.getOxygenLevel() < 8 && this.game.getOxygenLevel() + steps >= 8) {
-        tr.temperature = (tr.temperature ?? 0) + 1;
-      }
-    }
-
-    if (tr.temperature !== undefined) {
-      const availableSteps = Math.floor((constants.MAX_TEMPERATURE - this.game.getTemperature()) / 2);
-      const steps = Math.min(availableSteps, tr.temperature);
-      total = total + steps;
-      if (this.game.getTemperature() < 0 && this.game.getTemperature() + (steps * 2) >= 0) {
-        tr.oceans = (tr.oceans ?? 0) + 1;
-      }
-    }
-
-    if (tr.oceans !== undefined) {
-      const availableSteps = constants.MAX_OCEAN_TILES - this.game.board.getOceansOnBoard();
-      const steps = Math.min(availableSteps, tr.oceans);
-      total = total + steps;
-    }
-
-    if (tr.venus !== undefined) {
-      const availableSteps = Math.floor((constants.MAX_VENUS_SCALE - this.game.getVenusScaleLevel()) / 2);
-      const steps = Math.min(availableSteps, tr.venus);
-      total = total + steps;
-      if (this.game.getVenusScaleLevel() < 16 && this.game.getVenusScaleLevel() + (steps * 2) >= 16) {
-        tr.tr = (tr.tr ?? 0) + 1;
-      }
-    }
-
-    // MoonExpansion.ifMoon(this.game, (moonData) => {
-    //   if (tr.moonColony !== undefined) {
-    //     const availableSteps = constants.MAXIMUM_COLONY_RATE - moonData.colonyRate;
-    //     total = total + Math.min(availableSteps, tr.moonColony);
-    //   }
-
-    //   if (tr.moonMining !== undefined) {
-    //     const availableSteps = constants.MAXIMUM_MINING_RATE - moonData.miningRate;
-    //     total = total + Math.min(availableSteps, tr.moonMining);
-    //   }
-
-    //   if (tr.moonLogistics !== undefined) {
-    //     const availableSteps = constants.MAXIMUM_LOGISTICS_RATE - moonData.logisticRate;
-    //     total = total + Math.min(availableSteps, tr.moonLogistics);
-    //   }
-
-    total += tr.tr ?? 0;
-
-    return total;
   }
 
   private getStandardProjects(): Array<StandardProjectCard> {
